Change the on_stop method to run_clean in workchains
These workchains were originally developed within a different
version of Plum where the on_stop method of the Process class
was only called when the process was terminated, either successfully
or exceptionally. It was therefore really convenient to override it
to use for the cleaning of calculation directories at the end of the
WorkChain, without having to add a separate function to the spec.outline

However, in the current version of plum that is used in the latest
AiiDA release, the on_stop method of the Process class is called at every
state transition, causing the cleaning code to be called not just at the
end of the workchain but between every step. This is clearly not the
intended behavior.