Add implementation to deal with submission failure to PwBaseWorkChain
The implementation is extremely simple. After a submission failure
it sets the context flag has_submission_failed to True and simply
continues in the loop, relaunching a new PwCalculation with the
same inputs. If a calculation fails due to submission failure and
the flag is set the workchain is aborted. The has_submission_failed
is reset if a calculation is in FAILED state, which means that at
least submission was successful.

In short, only when two consecutive calculations end up in the
SUBMISSIONFAILED status, is the workchain aborted