Improve the flow and logic of the PhBaseCalculation skeleton
Simplified and improved the logical flow of the inspect_ph method
and use of context variables to indicate the status of the previously
run calculation, such as those for submission and unexpected calculation
failures. These flags are now set and reset in as little places as
possible, to keep reasoning about the flow easy. Hopefully this should
also make it easy for workchains that inherit from this one, to simply
override the _handle_* methods to do custom error handling.