Further abstract the main flow and error handling
By generalizing the _handle_calculation_failure to simply loop over
the calculation_failure_modes member of the workchain, which is an
ordered dictionary of error labels and the corresponding class methods,
we allow for the workchain to be easily extended by external plugins.
By simply adding more error handling methods to that dictionary, the
same _handle_calculation_error flow can be used, while still allowing
for more potential failure modes of PwCalculations to be handled.