Add dictionary with default QE input parameter values
In many error handlers the input parameters have to be
adapted based on the values of the previous calculation.
In the case that the parameter of interest was not
explicitly set, we know that the calculation will have
used the default value for that input parameter. This
dictionary defines these default pw.x values.