Fix bug in the generation of pseudo inputs in PwBaseWorkChain
Legacy calculations employ various 'use' methods to couple input nodes
to the calculation node. These functions allow for the definition of an
additional parameter. The 'use_pseudo' is such a use method where the kind
name with which the UpfData is to be associated is passed as the additional
parameter.

In the new system that requires the creation of a calculation process in order
to be able to run it, the inputs have to be passed in all at once as keyword
arguments. In order to keep backwards compatibility and maintain support for
additional parameters, inputs that require the additional parameter can pass
a dictionary where the key is the value of the additional parameter and the
value is the value of the input.

Taking an example for the 'use_pseudo' when we want to set the Ti UpfData
for the kinds Ti1 and Ti2, we would have called the following

	upf = UpfData()
	calc = PwCalculation()
	calc.use_pseudo(upf, kind=['Ti1', 'Ti2'])

In the new system this would become

	process = PwCalculation.process()
	submit(process, 'pseudo': {"('Ti1', 'Ti2')": upf})

The key of the dictionary is now the tuple of kind names that need to be associated
with that particular UPF. Note that we had to use a tuple, because keys of
dictionaries need to be immutable and hashable.