Add band data files to retrieve list when calculation is set to 'bands'
When a PwCalculation has a parameters node with the control calculation
set to 'bands', the user clearly wants to perform a band calculation and
would expect a bands data node to be parsed and attached in the output.
Currently, one had to specify the 'also_bands' key in the settings input
node. Now we also check the calculation mode in the parameters input node
and if set to 'bands' will add the band files to the retrieve list.