Add 'final_scf' boolean option to PwRelaxWorkchain
For some version of Quantum ESPRESSO the final charge density of a
completed 'vc-relax' calculation cannot be safely used for a restart.
In that case, one wants to perform a separate scf calculation on
the final relaxed structure. The workchain now provides an option that
is False by default, to run an scf calculation for the final relaxed
structure. The final remote folder and structure that will be attached
to the workchain can then safely be used for restart calculations.