Add the optional input 'group' to the PwBandsWorkChain and PwRelaxWorkChain
The two workchains now take an optional input called 'group' that when
specified will be used to store the final output_band node and final
completed PwCalculation, respectively. It currently has to be specified
as a Str data node which is not ideal, but the only possible solution.

Ideally, we would be able to pass the Group itself directly as an input
but a Group node is not a valid Data node and can therefore not be stored
officially as an input node and rightfully so. The solution would be to
mark it as 'non storable' which currently is accomplished by prefixing
the input key with an underscore. This behavior will be deprecated in the
near future, but would work if it weren't for input_group inputs where
this approach breaks down. The current input_group implementation requires
that all elements in it are storable, procluding these non storable
inputs with a leading underscore. This would prevent the group input to
be passed to the PwRelaxWorkChain, via the PwBandsWorkChain.

In the future, these two problems should be solved with the arrival of
PortNameSpaces and the expose_input functionality in plumpy v0.9* and
the addition of the 'non_db' keyword to the ProcessSpec.input() call.