Change the PwParser to always parse bands and create output_band node
A new feature was added to AiiDA core, that allows a Calculation to
specify a list of files or file patterns that are to be retrieved after
the calculation has finished successfully, but that are only to kept
during the parsing procedure and should not be permanently stored in
the retrieved FolderData node. This was implemented exactly for the
PwCalculation, which know allows to parse the band eigenvalues and
occupations at all k-points, without having to retrieve and store all
the k-point eigenvalue.xml files. This saves a lot of space in the
repository that would otherwise be taken up by the relatively useless
files.

The k-point eigenvalue.xml files are now by default always retrieved
in the temporary folder and are parsed by default. To parse these
we need the data-file.xml, which is also retrieved and stored by
default, but I have now added an explicit check that it has been
retrieved successfully or the parsing will fail.