Add a warning log message when 'also_bands' is used as setting
The 'also_bands' setting was deprecated as the bands are now
parsed by default and so it was replaced by the 'no_bands' key
to be able to return to the old behavior. If a user still passes
the 'also_bands' key, a log warning will be attached to the
calculation node to warn the user of the deprecation and the
fact that the eigenvalue.xml files at each k-point are no longer
saved explicitly in the repository.