Improve the generation of the options['resources'] dictionary
The JobCalculation process expects an input 'options' which should
have a key 'resources' which will contain all the relevant settings
for the scheduler. Exactly which keys are required or allowed is
scheduler dependent. Here we improved the generation of this resources
dictionary based on the automatic parallelization suggestion based
on the scheduler that is defined for the computer of the code that
is being run in the workchain.