Use deep copies in PwBaseWorkChain to make sure changes are local
The use of the base dictionary of inputs in the context to generate
a start for the inputs dictionary for a specific PwCalculation by
copying it, works but by not using a deepcopy and settings that
are changed in the local copy will be reflected in the base dict
in the context. This means that if one uses it another time, expecting
the inputs to reflect the original workchain inputs, some of the
settings that were set during earlier steps in the workchain may
linger. To prevent this we use a deepcopy which will not use a
reference of any kind.