Fix bug in parser_opts of PwParser and pass to parse_pw_text_output
The PwParser tried to retrieve the 'parser_options' from the settings
input node, but failed to convert the key to uppercase. Since during
calculation submission preparation all keys in the dictionaries are
converted to uppercase to ensure homogeneity, during parsing the
key never seemed to be specified. Converting the key to uppercase
now properly finds the parser options if specified.

The parser options are now also passed to the parse_pw_text_output
function.