Parse atomic occupations when set in the 'parser_options' of the settings
We provide a new key for the 'parser_options' key in the settings input.
When the user sets 'atomic_occupations' to True, the parser will attempt
to parse the atomic occupations which will be stored in a separate ouput
ParameterData node labeled 'output_atomic_occupations'. Note that this
information is only printed for DFT+U calculations.