Cleanup interface and scripts for PwRelax and PwBase
PwRelaxWorkChain had the settings input as required which is not
necessary. This is now optional. The input validation is now also
improved to make sure to create the 'CONTROL' card in the input
parameters if it is not there. This allows for the clean up of
the testing scripts that no longer need to explicitly define
an empty 'CONTROL' dictionary. The PwBaseWorkChain will by
default select an 'scf' calculation if no specific calculation
is specified