Fix bug when passing both final_scf and clean_workdir to PwRelaxWorkDir
When one passes both final_scf and clean_workdir to the PwRelaxWorkChain
the final scf calculation would fail. This is because the clean_workdir
input was just funneled to the PwBaseWorkChain, each one of which would
therefore be cleaned as soon as it was finished, but the workchain that
would run the final scf calculation would depend on the remote data of
the previous workchain, which would already have been cleaned.

To fix this, the clean_workdir option should not be passed to the
PwBaseWorkChain but the PwRelaxWorkChain itself should take care of
cleaning all the calculations of its sub workchains at the end and
only at the end of its own execution