Use Process on_destroy method for clean logic in workchains
The workchains PwRelax and PwBase provide the optional input
'clean_workdir' to clean the work directories of the calculations
they called during execution at the end of the workchain.
This was accomplished with a step in the outline called 'clean'
that was the final step of the outline. However, if for whatever
reason, the workchain was aborted, the final step would never be
executed and the calculation work directories would never be cleaned.

To fix this, we override the 'on_destroy' method of the parent
class of the WorkChain and add the cleaning logic there. This method
is guaranteed to always be called at the end of a Process execution
when it is being destroyed. This guarantees that the clean logic
is always executed. The only additional check we had to put in place
is to check whether the workchain itself had reached a terminal state
by calling 'has_finished()' to prevent the logic from being called
after every daemon tick.