First implementation of the AbstractBaseWorkChain
A big part of the machinery of the PwBaseWorkChain was dealing with
failure and convergence mechanisms that will be general to all sorts
of calculations and therefore workchains. It would be good to
abstract this functionality and have the PwBaseWorkChain extend it.
This way the core error handling can be reused for various base
workchains for different calculations