Remove calculation state check from parse_with_retrieved
The parsing function does not store any nodes as this is
the responsability of the exec manager which also calls
this function and should therefore be responsible for
ensuring that the calculation is put in the correct state
prior to parsing. Removing the checks from the plugins
also allows for easier testing of the parse function.