Renamed and moved BaseWorkChain to BaseRestartWorkChain
This particular base workchain implements the notion of 'restarting'
a calculation from a previous calculation that did not complete
successfully until it does or a maximum number of iterations is
exceeded. This concept appears often, for example in QE pw.x
calculations that can restart if convergence was not reached.