Turn ErrorHandlers into namedtuples with a priority
Wrapping the error handling methods for a WorkChain allows one
to add a 'priority' field, based on which the order of execution
can be defined by the user in a more controlled way. The failure
handling method will load all the ErrorHandlers defined in the
error handler entry point defined for the WorkChain class and
will sort them in reverse according to their priority value.