WorkChain error handlers now defined with register_error_handler decorator
Implement a decorator called @register_error_handler that can be applied
to a function, while taking a WorkChain class and a priority integer.
The function needs to take the arguments 'self' and 'calculation' which
when called will refer to the instance of the workchain that is being
run and the calculation that is currently being handled, respectively.
An example then would look like the following

	@register_error_handler(PwBaseWorkChain, 100)
	def _handle_error_convergence_not_reached(self, calculation)
	"""
	This function will handle the failed calculation for the
	PwBaseWorkChain if the reason for the failure was that convergence
	was not reached
	"""
	if 'convergence not reached' in calculation.errors:
		do something

The decorator will register the unbound method as a classmethod to the
workchain class provided in the decorator and set its priority to the
value defined. When the workchain is run and a calculation is encountered
in the FAILED state, the _handle_calculation_failure will loop over all
registered error handlers, sorted from high priority to low. A user can
register additional handlers by setting the _error_handler_entry_point
to an entry point category. Adding an entry point to this category that
points to a file with additional decorated functions, will make that upon
construction of the WorkChain the file is imported and therefore all
decorated functions will be registered as error handlers for the workchain.