Delegate max_iterations and clean_workdir to BaseRestartWorkChain
The logic that handles the max_iterations and clean_workdir inputs
is completely contained within the BaseRestartWorkChain. It therefore
also makes sense to define these inputs in the spec of the same
base workchain, such that an inheriting workchain does not have to
and can also not forget.

Moreover, the various variables used during the logical flow of the
BaseRestartWorkChain stored in the context, such as is_finished
will be set in the 'setup' method, that should be called in the
outline. These variables should only be used by the base restart
workchain. If the subclassing workchain decides to extend the setup
outline method, the parent's class method should be also called.