Make sure changes to self.ctx.inputs are used properly
The idea is to at setup define an AttributeDict in the context
self.ctx.inputs_raw which will store all the inputs necessary
for the calculations based on the input nodes to the workchain.
Changes that need to be kept for all calculations, e.g. options
and resource settings determined by an initialization calculation
when automatica parallelization was selected, can be added here.

After that is done, a deep copy will be stored at self.ctx.inputs.
This is the dictionary that will be used to create the inputs for
each calculation that will be launched. Error handlers should
update this dictionary with new inputs. If previous changes need
to be reset, the nodes from self.ctx.inputs_raw can be copied over.