Log INFO message for each error handling plugin loaded in BaseRestartWorkChain
For debugging purposes it is useful for the user to be able to see which
error handling entry points were loaded in the constructor of the
BaseRestartWorkChain. If certain error handlers were not triggered but
were expected by the user, at least it can be checked whether at
construction time they were loaded.