Use `parent_folder` instead of `parent_calc` for PhBaseWorkChain
The current PhBaseWorkChain uses parent_calc as an input for the restart
calculation, however, because Calculation nodes could originally not be
inputs to WorkCalculation nodes, a FrozenDict is wrapped around it. This
complicates the provenance graph a bit. Since the PhCalculation also
accepts a RemoteFolder as parent_folder it is better to have the
PhBaseWorkChain also accept this as input