Allow args of overridable option to be overriden
Certain predefined cli options are general enough that they can
be used multiple times for the same cli entry point. However, this
means then that the argument flags that are used should be different.
To allow this, the *args of the overridable_option should be also
overridable in the __call__. For example one can take the 'code'
overridable_option which defines:

	code = overridable_option(
    	'-c', '--code', type=click.STRING, required=True,
    	callback=validators.validate_code,
    	help='the label of the AiiDA code object to use'
	)

One can imagine a cli script, however, that requires two different
codes as input. It would then be impossible to use the option. By
allowing the args to be overwritten this is made possible. e.g.:

	@options.code('--ph', 'code_ph', help='code label referencing ph.x')
	@options.code('--pw', 'code_pw', help='code label referencing pw.x')