Implement optional binding of kwargs to callback function
Certain callbacks may need additional arguments passed by the user
upon calling of the overridable_option. To facilitate this, we allow
the passing of the callback_kwargs keyword to the __call__ of the
overridable_option class. If a callback is specified, these arguments
will be bound to it through functools.partial.

This new functionality is used for the parent_calc option which now
allows a user to pass a specific calculation entry point to provide
an explicit calculation class that the option expects