PwRelaxWorkChain: fix bug when adding final PwCalculation to group
When the 'group' input is passed to the workchain, upon workchain
success, the last completed PwCalculation will be added to said group.
The PwCalculation was retrieved from the final PwBaseWorkChain, either
the from the last relax cycle or the final scf run, from its ouput
parameters node, from which in turn through the .inp method the
PwCalculation was returned. However, since the PwBaseWorkChain also
return the output_parameters ParameterData node, it will no longer
have just a single parent node with the link label 'output_parameters'.
The .inp method will then randomly select one, which sometimes resulted
in the WorkCalculation node being returned instead of the desired
PwCalculation node. To fix this ambiguity we use the get_inputs()
method instead, explicitly stating that we want the parent node with
node type PwCalculation.