Add parsing of electronic and ionic dipole for raw pw output
When the boolean switch 'lelfield' is defined in the CONTROL card of
the input parameters of a PwCalculation, the electronic and ionic
dipoles will be printed at the end of every scf iteration, both the
average over the cell as well as projected on the cartesian axes.

Add the parsing of these values for the final scf iteration of each
trajectory step under the following keys:

  * electronic_dipole_cell_average
  * electronic_dipole_cartesian_axes
  * ionic_dipole_cell_average
  * ionic_dipole_cartesian_axes

Each of these fields will be a list with as many entries as there were
scf steps in the trajectory.