PwBandsWorkChain: make 'relax' inputs and relaxation optional
The relaxation of the input structure was made optional by defining
a 'skip_relax' optional boolean input. However, the 'relax' input
group was always required. We remove the 'skip_relax' input and just
determine whether the input structure needs to be relaxed based on
whether the 'relax' input group was specified in the workchain inputs
or not.