PwBandsWorkChain: use primitive structure for kpoints mesh generation
The kpoints_distance value was used to generate a kpoints mesh for the
scf step, when specified, but it did so on the input structure. However,
the structure that was actually used for the calculation, was the one
returned by Seekpath, which may have changed, since Seekpath tries to
reduce the structure to the primitive cell. To ensure that both the
kpoints mesh for the scf and the kpoints path (determined by Seekpath)
for the bands calculation are commensurate with the structure that is
used, we use the primitive structure returned by the Seekpath analysis