Update the state changing callback on_destroy to on_terminated
We used the on_destroy handle to action the cleaning of the work
directories if asked by the user. This has now changed to on_terminated.
This has the added advantage that upon the call of this method the
process is guaranteed to be in a terminal state, unlike with on_destroy,
which is why we no longer have to check manually that the state of the
workcalculation is indeed terminal