Correctly implement clean workdir functionality for all workchains
For workchains that launch multiple calculations or sub workchains
who might depend on the result of one another, have to defer the
cleaning of the work directories until the entire process is terminated.
This includes making sure not to pass clean_workdir=True to any
sub workchains and only cleaning the work directory for all called
descendants once the parent process is terminated. This is made easier
by a new Calculation property 'called_descendants' which will recursively
retrieve a list of all called descendant calculations. One then only has
to find the JobCalculation instances and attempt to call the clean
method on their remote folder node.