Do not use deepcopy on inputs dictionaries in workchains
If the copy operators of Nodes were to be implemented correctly, this
usage would result in duplicated entries to be generated in the database
which is not the intended behavior. Instead we keep one dictionary of
inputs and simply pass it to `prepare_process_inputs` which will wrap
bare dictionaries that are intended to be ParameterData objects in the
ParameterData class. With this change, one does have to remember that
when reusing the inputs dictionary for reuse, that potential changes in
a previous iteration of the calculation or sub workchain will be have
to be removed if necessary