Move tests and data outside main aiida_quantumespresso module (#154)
The test data is heavy and should not be included in the actual
package, since it is not needed and makes the package unnecessarily
big. We move the tests to the main level of the repository and
remove the lines from the `MANIFEST` such that they won't be included
in the distributions. For the tests on Travis to still work I did have to adjust
the script to copy the tests folder from the repository to the virtual
environment directory, otherwise the would not be found when the tests
would be run through `verdi devel` tests. When the repository is checked
out locally for development this step is not necessary as the tests can find
the tests data folder directly in the repository.