Implement the new expose port functionality in workchains
A recent update in aiida-core made it possible to properly expose
the input ports of a workchain in a wrapping workchain. The whole
exposed process class can now also be made to be optional, which
means that if that namespace is left empty when the workchain is
instantiated, the validation will not be triggered.

This mechanism is for example used in the PwBandsWorkChain, which
will expose the inputs of the PwRelaxWorkChain, but make the space
optional. This means that the presence of the `relax` namespace
in the inputs can now be used as a switch to determine whether the
user wants the structure to be relaxed.

Additionally, we provide support for defining smearing, starting
magnetization and Hubbard parameters through the command line
interface launch scripts, including the corresponding validation
functions.