Update the version to v2.1.0 and add parser note to docs
From this point on, the parser version will be tied to the version of
the package. In addition, it will be added to the output parameters
node of the calculation under the key `parser_version`. A section is
added to the docs of pw.x to mention this change and how to retrieve
the parser version of a completed calculation.