Fix the parser tests to pass retrieve_temporary_folder correctly
In the current version of aiida-core, the retrievee_temporary_folder, which
contains files that are supposed to be retrieved and be made available only
for the parsing process, is no longer a FolderData but simply an absolute
path to the directory containing the retrieved files. The test archives are
expected to have these temporary files in the repository directory of the
retrieved node. When new test archives are generated, the test generator
should make sure that when the `retrieved_temporary_list` key is specified,
that the corresponding files are copied in the retrieved folder location
of the archive, or the parser test will fail.