Exclude `clean_workdir` in expose inputs of wrapping workchains (#184)
Some of the sub workchains launched by the wrapping workchain, may
depend on the remote folder contents on one of the calculations ran
by another sub workchain. Letting the sub workchains take care of
cleaning the workdirs of the launched calculation may therefore lead
to premature cleaning. Instead, cleaning of workdir is deferred to
the wrapping workchain, who at the end of its steps will clean the
directories of all calculations that were called by any of the child
processes called downstream.