Implement parsing of XML files from the new XSD schema files
This adapts the calculation classes and parsers for the `cp.x`, `pw.x`
and `neb.x` executables, which are affected, to be able to parse from
both the old XML output file and the new XML which will be parseable
with the use of an XSD schema files. The old format will be fully
removed from Quantum ESPRESSO 7.0, but for Quantum ESPRESSO  6.* both
versions can be used depending on the way the code was compiled so the
calculation and parsers plugins need to be able to deal with both. The
change in this commit performs the initial infrastructure to retrieve
either file and based on which one is found, trigger the corresponding
XML parsing path.