Check that PwBaseWorkChain has output_structure in PwRelaxWorkChain (#194)
The inspect_workchain did check whether the PwBaseWorkChain was marked
as finished_ok, however, it assumed that in that case the output
structure node would be there. Given that the output_structure is not
a required output node for the PwBaseWorkChain, this assumption could
lead to an exception if the output node is not present, for whatever
reason.