Improve kpoint mesh selection algorithm in `create_kpoints_from_distance` (#203)
The `create_kpoints_from_distance` workfunction is used by among others
the pw.x workchains to generate a suitable kpoint mesh for a given
structure with a minimal kpoint distance. However, for certain oddly
defined cell structures, this can generate meshes that result into
problems in pw.x. For example for certain cells with equal length
vectors in real space, the cell in reciprocal space can be non isotropic
causing a non-isotropic mesh to be generated. This leads to problems with
the symmetry operations leading to more kpoints being calculation than
necessary. We solve this particular problem by comparing the proposed
kpoint mesh to the real space cell. If the real space cell has vectors of
equal length but the mesh is non-isotropic, the mesh is made to be isotropic
by taking the highest number of kpoints and choosing that in all directions.