Update `xmlschema` to 1.0.6 and improve XML parsing
The parsing from the new XML based on the schema has been improved in
terms of parsing more information that is now available from the XML as
well as improving the naming of certain keys. Since this will make the
produced output nodes different from the older version, a switch is
introduced that allows a user to force the parser to use the old naming
for the keys. Properties whose parsing has been improved or changed:

 * Bands
 * Occupations
 * K-points
 * Eigenvalues
 * Monkhorst Pack grid and offset
 * Lattice symmetries added
 * Smearing
 * Parse `status` tag
 * Parse `output/convergence_info` tag
 * Parse `nrot` and `nsym` in `output/symmetries` for consistency checks
 * Parse in each symmetry `info/@time_reversal`, formerly `t_rev`
 * Berry phase data