BaseRestartWorkChain: fix error handlers not being loaded during instance load. (#218)
The loading of `error_handlers` through the plugin entry point was
only being performed on construction. This means that when a
workchain was being loaded from instance state, which basically
means when it is run over the daemon, the error handler added
through the entry point would not be configured. Here we override
the `load_instance_state` to trigger the logic upon loading as well.