Turn `workfunctions` into `calcfunctions`
The `workfunctions` used in the various workchains create new data from
the inputs they receive and should therefore, according to the new
provenance paradigm of `aiida-core==1.0.0`, should be `calcfunctions`
instead. In this commit, the decorators are adapted and the functions
are moved to individual files. Given that the source code of the file in
which the functions are defined, is copied to the repository for each
instance of the function that is run, one has to make sure that it only
contains the code necessary. Putting it in the file of the workchain
itself will copy, unnecessarily also the source code of the workchain
each time. Until this inefficiency is automatically optimized within
`aiida-core` itself, we should keep process functions in separate files.