Replace the deprecated `_input_subfolder` with literal constant
The `_input_subfolder` variable, which used to evaluate to `raw_input`,
has been removed in `aiida-cor==1.0.0b1` and has been moved to the
repository utility that is a property of each node. Since the logic has
changed significantly with respect to `JobCalculations` which are now
`CalcJobs` there is no one-to-one replacement variable so we simply use
the literal value here. In any case, this is used in the immigrant which
is fundamentally broken and needs to be redesigned.