Bare essentials to make `PwCalculation` run in py2 and py3
For the most part, the problem had to do with writing unicode strings to
a binary handle, which of course fails in python 3. We now write to a
handle that will write encoded text and make sure the string it receives
is unicode in both versions.

Additionally the `translate` string method has changed interface in py3
which was used in the XML parser.